pyproject.toml
[project]
name = "segmentation-model"
version = "0.1.0"
description = "DeepLabV3+ ground surface segmentation using the RUGD dataset"
requires-python = ">=3.11,<3.14"
dependencies = [
    # Deep learning
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "segmentation-models-pytorch>=0.3.3",
    # Data & augmentation
    "numpy>=1.26.0",
    "pillow>=10.2.0",
    "opencv-python-headless>=4.9.0",  # headless: no GUI deps (safe for HPC)
    "albumentations>=1.3.1",
    # Config & utilities
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    # Logging & visualization
    "tensorboard>=2.16.0",
    "matplotlib>=3.8.0",
    "wandb>=0.17.0",
]

[project.optional-dependencies]
# Install on local dev machine only
dev = [
    "ruff>=0.3.0",
    "ipykernel>=6.29.0",
]

[tool.uv]
# This is a scripts project — uv should not try to install it as a package.
package = false
# On HPC, install PyTorch from the CUDA-specific index.
# Run: uv pip install torch torchvision --extra-index-url https://download.pytorch.org/whl/cu121
# (replace cu121 with your cluster's CUDA version, e.g. cu118, cu124)

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

